RDKAAMP-3636 "address new aamp-2411 sprint compilation warnings from xcode static code analysis" which will be used to clean a handful of new issues that have crept in
Reason for change: address new compilation warnings
Test Procedure: light regression (absolutely no impact expected)
Risks: None

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_2

+++ priv_aamp.cpp
@@ -11141,7 +11141,7 @@ int PrivateInstanceAAMP::GetTextTrack()
 			{
 				if (it->instreamId == trackId)
 				{
-					idx = std::distance(tracks.begin(), it);
+					idx = static_cast<int>( std::distance(tracks.begin(), it) );
 				}
 			}
 		}
